Move window dependencies out of streamer
No tests!?
